// app/types/team.ts
export interface PublicTeamMember {
    Name: string | null;
    College: string | null;
    Course: string | null;
    "Branch ": string | null;
    "Year of Study": number | null;
    City: string | null;
    State: string | null;
    Themes: string | null;
    "Team Name": string | null;
  }
  
  export interface TeamMember {
    Timestamp: number | null;
    Name: string | null;
    Email: string | null;
    "Phone number": number | null;
    Gender: string | null;
    College: string | null;
    Course: string | null;
    "Branch ": string | null;
    "Year of Study": number | null;
    City: string | null;
    State: string | null;
    Themes: string | null;
    "Team Size": number | null;
    "Team Name": string | null;
    "Upload Payment Screenshot (Individual)": string | null;
    "I confirm that I am registering individually and have read the instructions.": string | null;
    "Payment status": string | null;
  }